Add line-of-sight for enemy mobs (#10)
This commit implements a simple field-of-vision implementation.
If the player is within the enemy line of sight, then it moves.
Instead of using Area2D etc., I just used high-school geometry
to compute if a point (player pos) is within the radius of a circle
(defined by enemy pos and line of sight as radius).