Add init experiments for death animation
I really don't understand how I was able to make it work. But my
breakthrough is to instead add the death animation in the remove()
function. I'm not sure why it doesn't work in take_damage...

In retrospect, it actually makes sense, because you want to have the
death animation when the enemy is about to be removed.