Add heads-up display (HUD) (#21)
Resolves #7 

As it turns out, creating the HUD is quite easy. I really don't need to make it super beautiful. As long as it shows the Level,
my HP, and current Dmg, then I'm actually goods. Not sure where I should put the weapon. Maybe in the upper right? That's an issue for future me then!