Attempt fix for orphaned enemies (#46)
Resolves #46

So what happens is that whenever we generate a level, we add the enemies
into the Main Node (Game). However, when we remove() it (either by
pressing restart or by advancing to the next level), we actually just
clear the TileMap, but not the enemies that were added into the level.

What this fix does is that it tries to delete the enemies generated for
that instance.

Q: Why didn't the build_level delete it?
A: Because the build_level()'s clear happens to the *new* instance, not
on the previous one. So it's not really clearing anything. I think the
original idea is to always reuse that instance. For us, since we're
using a Factory pattern, we just delete the instance and create a new
one.